to create zip, run "make zip"; to test zip, run "make testzip"

+++ src/simpledu_args.c
@@ -55,7 +55,7 @@ int simpledu_args_ctor(simpledu_args_t *p, int argc, char *argv[]){
             case 'l': p->count_links   = true; break;
             case 'S': p->separate_dirs = true; break;
             case 'B': if(sscanf(optarg, "%llu" , &p->block_size  ) != 1) return EXIT_FAILURE; break;
-            case 'd': if(sscanf(optarg, "%hu"  , &p->max_depth   ) != 1) return EXIT_FAILURE; break;
+            case 'd': if(sscanf(optarg, "%hd"  , &p->max_depth   ) != 1) return EXIT_FAILURE; break;
             case 'f': if(sscanf(optarg, "%d"   , &p->log_filedes ) != 1) return EXIT_FAILURE; break;
             case 'p': if(sscanf(optarg, "%d"   , &p->pipe_filedes) != 1) return EXIT_FAILURE; break;
             case 's': if(sscanf(optarg, "%lldd", &p->start_time  ) != 1) return EXIT_FAILURE; break;
@@ -74,11 +74,11 @@ int simpledu_args_ctor(simpledu_args_t *p, int argc, char *argv[]){
     if(p->filesc != 0) {
         p->files  = malloc(p->filesc*sizeof(char *));
         if(p->files == NULL) return EXIT_FAILURE;
-        for(int i = 0; i < p->filesc; ++i){
+        for(size_t i = 0; i < p->filesc; ++i){
             const char *s = argv[optind+i];
             char *str = malloc((1+strlen(s)*sizeof(char)));
             if(str == NULL){
-                for(int j = 0; j < i; ++j) free(p->files[i]);
+                for(size_t j = 0; j < i; ++j) free(p->files[i]);
                 free(p->files);
                 return EXIT_FAILURE;
             }
@@ -104,7 +104,7 @@ int simpledu_args_ctor(simpledu_args_t *p, int argc, char *argv[]){
 
 int simpledu_args_dtor(simpledu_args_t *p){
     if(p == NULL) return EXIT_SUCCESS;
-    for(int i = 0; i < p->filesc; ++i){
+    for(size_t i = 0; i < p->filesc; ++i){
         free(p->files[i]);
         p->files[i] = NULL;
     }
@@ -196,7 +196,7 @@ int simpledu_args_toargv(const simpledu_args_t *p, char ***pargv){
     { sprintf(buf, "--pipe-filedes=%d", p->pipe_filedes); strcpy(argv[i++] = malloc((strlen(buf)+1)*sizeof(char)), buf); } //pipe_filedes
     { sprintf(buf, "--start-time=%lld", p->start_time  ); strcpy(argv[i++] = malloc((strlen(buf)+1)*sizeof(char)), buf); } //pipe_filedes
     
-    for(int j = 0; j < p->filesc; ++i, ++j){
+    for(size_t j = 0; j < p->filesc; ++i, ++j){
         argv[i] = malloc((strlen(p->files[j])+1)*sizeof(char));
         strcpy(argv[i], p->files[j]);
     }
